ci(commitlint): 🐛 Fix job failing due to unknown range
The container in the GitHub Actions workflow doesn't know the names of
all branches available in the repository.
We had to use a workaround using a variable provided by GitHub actions
to get a count of commits messages in the target branch.
It was needed to specify the range for commitlint, to how many commits
it is supposed to check.

Co-authored-by: lax0248917 <zach_a_king@outlook.com>