docs:  📝 Initial project documentation
➕  (./README.md)
Add a new documentation file with initial sections to make an entry
project presentation. It should be easy to read and encouraging aspiring
contributors to get involved.

➕ (./CONTRIBUTING.md)
Add a new documentation file with initial instructions on developing steps.
It should shortly describe the set-up development flow for the aspiring
contributors to navigate through the project easily.

Co-authored-by: lax0248917 <zach_a_king@outlook.com>